// src/congress-ingestion/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, KvCache, Logger, MRNObject, ServiceStub, SmartBucket, SqlDatabase, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  ADMIN: ServiceStub<import('../admin/index.js').default>;
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  API_CACHE: KvCache;
  API_GATEWAY: ServiceStub<import('../api-gateway/index.js').default>;
  AUTH_SERVICE: ServiceStub<import('../auth-service/index.js').default>;
  BILL_CHAT: ServiceStub<import('../bill-chat/index.js').default>;
  BILL_STORAGE: ServiceStub<import('../bill-storage/index.js').default>;
  BILLS_BUCKET: SmartBucket;
  HAKIVO_DB: SqlDatabase;
  logger: Logger;
  mem: KvCache;
  NEWS_AGGREGATOR: ServiceStub<import('../news-aggregator/index.js').default>;
  PODCAST_GENERATOR: ServiceStub<import('../podcast-generator/index.js').default>;
  REPRESENTATIVE_LOOKUP: ServiceStub<import('../representative-lookup/index.js').default>;
  SESSION_CACHE: KvCache;
  tracer: Tracer;
  VOICE_AGENT: ServiceStub<import('../voice-agent/index.js').default>;
}